moving channel switcher to its own component and removing dead code
