+++ kolibri/plugins/learn/assets/src/vue/toolbar/index.vue
@@ -2,15 +2,8 @@
 
   <div v-bind:class="['toolbar', displayToolbar ? 'toolbar-hide' : '']" v-show='!searchOpen'>
     <breadcrumbs class="breadcrumbs"></breadcrumbs>
-    <label for="chan-select" class="visuallyhidden">{{ $tr('switchChannels') }}</label>
-    <select
-      name="chan-select"
-      id="chan-select"
-      class="chan-select"
-      v-model="currentChannel">
-      <option v-for="channel in channelList" :value="channel.id">{{ channel.name }}</option>
-    </select>
-    <search-button @scrolling="handleScroll" class='search-btn'></search-button>
+    <channel-switcher></channel-switcher>
+    <search-button @scrolling="handleScroll" class="search-btn"></search-button>
   </div>
 
 </template>
@@ -18,97 +11,33 @@
 
 <script>
 
-  const constants = require('../../state/constants');
-  const getters = require('../../state/getters');
-  const PageModes = constants.PageModes;
-
   module.exports = {
-
-    $trNameSpace: 'learnToolbar',
-    $trs: {
-      switchChannels: 'Switch Channels',
-    },
-
     data: () => ({
       currScrollTop: 0,
       lastScrollTop: 0,
       delta: 5,
       displayToolbar: false,
-      more: false,
     }),
     components: {
       'search-widget': require('../search-widget'),
       'search-button': require('../search-widget/search-button'),
       'breadcrumbs': require('../breadcrumbs'),
-    },
-    computed: {
-      /*
-      * Get and set the current channel ID.
-      */
-      currentChannel: {
-        get() {
-          return this.currentChannelGetter;
-        },
-        set(newChannelId, oldChannelId) {
-          if (newChannelId !== oldChannelId) {
-            this.switchChannel(newChannelId);
-          }
-        },
-      },
+      'channel-switcher': require('./channel-switcher'),
     },
     methods: {
       handleScroll(position) {
         this.position = position;
         this.currScrollTop = position.scrollTop;
-
         if (Math.abs(this.lastScrollTop - this.currScrollTop) <= this.delta) {
           return;
         }
-
-        this.more = false;
-
-        if (this.currScrollTop > this.lastScrollTop) {
-          this.displayToolbar = true;
-        } else {
-          this.displayToolbar = false;
-        }
+        this.displayToolbar = this.currScrollTop > this.lastScrollTop;
         this.lastScrollTop = this.currScrollTop;
-      },
-      switchChannel(channelId) {
-        let rootPage;
-        this.more = false;
-        if (this.pageMode === PageModes.EXPLORE) {
-          rootPage = constants.PageNames.EXPLORE_CHANNEL;
-        } else {
-          rootPage = constants.PageNames.LEARN_CHANNEL;
-        }
-        this.clearSearch();
-        this.$router.go(
-          {
-            name: rootPage,
-            params: {
-              channel_id: channelId,
-            },
-          }
-        );
-      },
-      toggleMore() {
-        this.more = !this.more;
       },
     },
     vuex: {
       getters: {
-        rootTopicId: state => state.rootTopicId,
-        topic: state => state.pageState.topic,
-        isRoot: (state) => state.pageState.topic.id === state.rootTopicId,
-        pageMode: getters.pageMode,
-        pageName: state => state.pageName,
-        currentChannelGetter: state => state.currentChannel,
-        channelList: state => state.channelList,
         searchOpen: state => state.searchOpen,
-      },
-      actions: {
-        clearSearch: require('../../actions').clearSearch,
       },
     },
   };
@@ -143,16 +72,6 @@
     left: 120px
     @media screen and (max-width: $portrait-breakpoint)
       left: 1.3em
-
-  .chan-select
-    color: $core-text-annotation
-    font-size: 0.9rem
-    position: absolute
-    top: 0.5rem
-    right: 6em
-    @media screen and (max-width: $portrait-breakpoint)
-      transform: translateX(-50%)
-      left: 53%
 
   .search-btn
     position: absolute
